+++ src/main/java/com/thiagodeas/javastockmanager/repositories/OutboundRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.thiagodeas.javastockmanager.models.Outbound;
 import com.thiagodeas.javastockmanager.models.enums.OutboundReason;
 
-public interface OutboundRepository extends JpaRepository<Outbound, Integer>{
+public interface OutboundRepository extends JpaRepository<Outbound, Long>{
     List<Outbound> findByDateRange(LocalDateTime start, LocalDateTime end);
     List<Outbound> findByReason(OutboundReason reason);
     List<Outbound> findByDateRangeAndReason(LocalDateTime start, LocalDateTime end, OutboundReason reason);
